Issue #1 completed/resolved.  Added QB controller
Created the QB controller - cerated index and show
Confirmed that routes js added the resources

Edited the scripts for QB.SH.  It seems that created the database
collections to QB was not good.  mongo/express did not like that and
was not showing data when pulling a GET request.  Had to change the
collection name to qbs and now able to get handlebars etc working to
show the data i needed.  Updated the data files to remove weird
markings.

QB model/controller working without any issues! 1 of 7 tables completed